SVG 파일 분리, emotion 으로 변경
