+++ src/css/main.css
@@ -22,7 +22,6 @@
 @import 'base/MaintenancePage.css';
 @import 'base/TextTruncate.css';
 @import 'base/ExpandableBookList.css';
-@import 'base/AlertForNonSubscriber.css';
 @import 'intro/intro.css';
 @import 'home/index.css';
 @import 'search_result/index.css';
